fix(test): Handle case insensitivity.
The TSLint test's host says the file system is case-sensitive regardless
of whether it is or not. So more needs to be done to find any file
that's used for the add imports. (As path.resolves is called and it will
convert paths to lower case for case-insensitive file systems.)